# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ShopFlow API",
        "description": "Headless e-commerce order processing API.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local development"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Catalog"
                ],
                "summary": "List active products (paginated)",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 50)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{product}": {
            "get": {
                "tags": [
                    "Catalog"
                ],
                "summary": "Get a single product by slug",
                "operationId": "db00b523e69d5dbc09c6edad76374bdc",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details with variants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Category": {
                "properties": {
                    "slug": {
                        "type": "string",
                        "example": "t-shirts"
                    },
                    "name": {
                        "type": "string",
                        "example": "T-Shirts"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    }
                },
                "type": "object"
            },
            "Product": {
                "properties": {
                    "slug": {
                        "type": "string",
                        "example": "basic-tee"
                    },
                    "name": {
                        "type": "string",
                        "example": "Basic Tee"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "variants": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductVariant"
                        }
                    }
                },
                "type": "object"
            },
            "ProductVariant": {
                "properties": {
                    "sku": {
                        "type": "string",
                        "example": "SAA-478277"
                    },
                    "name": {
                        "type": "string",
                        "example": "S / white"
                    },
                    "price": {
                        "properties": {
                            "amount": {
                                "type": "integer",
                                "example": 45517
                            },
                            "currency": {
                                "type": "string",
                                "example": "USD"
                            },
                            "formatted": {
                                "type": "string",
                                "example": "455.17"
                            }
                        },
                        "type": "object"
                    },
                    "in_stock": {
                        "type": "boolean",
                        "example": true
                    },
                    "stock": {
                        "type": "integer",
                        "example": 27
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "error": {
                        "properties": {
                            "code": {
                                "type": "string",
                                "example": "not_found"
                            },
                            "message": {
                                "type": "string",
                                "example": "Resource not found."
                            },
                            "details": {
                                "type": "object",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 4
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 52
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}